# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: IDE Platform API — Replit-like workspace platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List current user's workspaces
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"

    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create a new workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
  /workspaces/{workspaceId}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get workspace by ID
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateWorkspace
      tags: [workspaces]
      summary: Update workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"

    delete:
      operationId: deleteWorkspace
      tags: [workspaces]
      summary: Delete workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /workspaces/{workspaceId}/start:
    post:
      operationId: startWorkspace
      tags: [workspaces]
      summary: Start workspace container
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
  /workspaces/{workspaceId}/stop:
    post:
      operationId: stopWorkspace
      tags: [workspaces]
      summary: Stop workspace container
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
  /workspaces/{workspaceId}/publish:
    post:
      operationId: publishWorkspace
      tags: [workspaces]
      summary: Publish workspace (permanent URL, always-on)
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
  /workspaces/{workspaceId}/unpublish:
    post:
      operationId: unpublishWorkspace
      tags: [workspaces]
      summary: Unpublish workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"

  # ─── FILES ──────────────────────────────────────────────────────────────────
...
  /workspaces/{workspaceId}/workflows:
    get:
      operationId: listWorkflows
      tags: [workflows]
      summary: List workflows
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workflow"

    post:
      operationId: createWorkflow
      tags: [workflows]
      summary: Create a workflow
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /workspaces/{workspaceId}/workflows/{workflowId}:
    patch:
      operationId: updateWorkflow
      tags: [workflows]
      summary: Update a workflow
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"

    delete:
      operationId: deleteWorkflow
      tags: [workflows]
      summary: Delete a workflow
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /workspaces/{workspaceId}/workflows/{workflowId}/start:
    post:
      operationId: startWorkflow
      tags: [workflows]
      summary: Start a workflow
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /workspaces/{workspaceId}/workflows/{workflowId}/stop:
    post:
      operationId: stopWorkflow
      tags: [workflows]
      summary: Stop a workflow
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /workspaces/{workspaceId}/workflows/{workflowId}/logs:
    get:
      operationId: getWorkflowLogs
      tags: [workflows]
      summary: Get recent log output for a workflow
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowLogs"

  # ─── DATABASES ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]

    # ─── AUTH ────────────────────────────────────────────────────────────────
...
    Workspace:
      type: object
      required: [id, name, slug, language, status, isPublished, userId, createdAt, lastActive]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        language:
          type: string
        status:
          type: string
          enum: [running, stopped, sleeping, error, starting]
        isPublished:
          type: boolean
        publishedUrl:
          type: ["string", "null"]
        devUrl:
          type: ["string", "null"]
        userId:
          type: string
        username:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        lastActive:
          type: string
          format: date-time
    WorkspaceInput:
      type: object
      required: [name, language]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: ["string", "null"]
        language:
          type: string
          enum: [nodejs, python, php, go, rust, java, ruby, dotnet, deno, bash, html, cpp]
    WorkspaceUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]

    # ─── FILES ───────────────────────────────────────────────────────────────
...
    Workflow:
      type: object
      required: [id, workspaceId, name, command, autoStart, status]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        command:
          type: string
        port:
          type: ["number", "null"]
        autoStart:
          type: boolean
        status:
          type: string
          enum: [running, stopped, error]
        pid:
          type: ["number", "null"]
    WorkflowInput:
      type: object
      required: [name, command]
      properties:
        name:
          type: string
          minLength: 1
        command:
          type: string
          minLength: 1
        port:
          type: ["number", "null"]
        autoStart:
          type: boolean
    WorkflowUpdate:
      type: object
      properties:
        name:
          type: string
        command:
          type: string
        port:
          type: ["number", "null"]
        autoStart:
          type: boolean
    WorkflowLogs:
      type: object
      required: [workflowId, lines]
      properties:
        workflowId:
          type: string
        lines:
          type: array
          items:
            type: string

    # ─── DATABASE ────────────────────────────────────────────────────────────
...