# lib/api-zod/src/generated/types/workflowLogs.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * IDE Platform API — Replit-like workspace platform
 * OpenAPI spec version: 0.1.0
 */

export interface WorkflowLogs {
  workflowId: string;
  lines: string[];
}